Fix(file names): Change for consistency
Renamed properly the account screen and edit to be more suitable than
the old names
