+++ app/src/androidTest/java/com/android/ootd/ui/account/AccountScreenTest.kt
@@ -1,51 +1,27 @@
 package com.android.ootd.ui.account
 
-/*
- * DISCLAIMER: This file was created/modified with the assistance of GitHub Copilot.
- * Copilot provided suggestions which were reviewed and adapted by the developers.
- */
-import android.content.Context
-import android.net.Uri
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextContains
-import androidx.compose.ui.test.hasAnyAncestor
-import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.isDisplayed
-import androidx.compose.ui.test.isToggleable
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import androidx.compose.ui.test.onAllNodesWithText
+import androidx.compose.ui.test.onAllNodesWithTag
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.performImeAction
-import androidx.compose.ui.test.performScrollTo
-import androidx.compose.ui.test.performTextClearance
-import androidx.compose.ui.test.performTextInput
 import androidx.credentials.CredentialManager
-import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.ootd.LocationProvider
 import com.android.ootd.model.account.Account
 import com.android.ootd.model.account.AccountRepository
 import com.android.ootd.model.authentication.AccountService
-import com.android.ootd.model.map.Location
-import com.android.ootd.model.map.LocationRepository
+import com.android.ootd.model.feed.FeedRepository
+import com.android.ootd.model.posts.OutfitPost
+import com.android.ootd.model.user.User
 import com.android.ootd.model.user.UserRepository
-import com.android.ootd.ui.map.LocationSelectionTestTags
 import com.android.ootd.ui.theme.OOTDTheme
-import com.google.android.gms.location.FusedLocationProviderClient
-import com.google.firebase.auth.FirebaseUser
 import io.mockk.clearAllMocks
 import io.mockk.coEvery
-import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
-import kotlinx.coroutines.flow.MutableStateFlow
 import org.junit.After
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -58,43 +34,64 @@ class AccountScreenTest {
   private lateinit var mockAccountService: AccountService
   private lateinit var mockAccountRepository: AccountRepository
   private lateinit var mockUserRepository: UserRepository
-  private lateinit var mockLocationRepository: LocationRepository
+  private lateinit var mockFeedRepository: FeedRepository
   private lateinit var mockCredentialManager: CredentialManager
-  private lateinit var mockFirebaseUser: FirebaseUser
-  private lateinit var viewModel: AccountViewModel
+  private lateinit var viewModel: AccountPageViewModel
 
-  private val userFlow = MutableStateFlow<FirebaseUser?>(null)
+  private var onEditAccountCalled = false
+  private var onPostClickCalledWithId: String? = null
+
+  private val testUser =
+      User(
+          uid = "test-uid",
+          username = "testuser",
+          profilePicture = "https://example.com/profile.jpg")
+
+  private val testAccount =
+      Account(
+          uid = "test-uid",
+          ownerId = "test-uid",
+          username = "testuser",
+          profilePicture = "https://example.com/profile.jpg",
+          friendUids = listOf("friend1", "friend2", "friend3"),
+          isPrivate = false)
+
+  private val testPosts =
+      listOf(
+          OutfitPost(
+              postUID = "post1",
+              name = "testuser",
+              ownerId = "test-uid",
+              outfitURL = "https://example.com/post1.jpg",
+              description = "First post"),
+          OutfitPost(
+              postUID = "post2",
+              name = "testuser",
+              ownerId = "test-uid",
+              outfitURL = "https://example.com/post2.jpg",
+              description = "Second post"),
+          OutfitPost(
+              postUID = "post3",
+              name = "testuser",
+              ownerId = "test-uid",
+              outfitURL = "https://example.com/post3.jpg",
+              description = "Third post"))
 
   @Before
   fun setup() {
     mockAccountService = mockk(relaxed = true)
     mockAccountRepository = mockk(relaxed = true)
     mockUserRepository = mockk(relaxed = true)
-    mockLocationRepository = mockk(relaxed = true)
+    mockFeedRepository = mockk(relaxed = true)
     mockCredentialManager = mockk(relaxed = true)
-    mockFirebaseUser = mockk(relaxed = true)
 
-    every { mockAccountService.currentUser } returns userFlow
-    every { mockFirebaseUser.uid } returns "test-uid"
-    every { mockFirebaseUser.email } returns "user1@google.com"
-    every { mockFirebaseUser.photoUrl } returns null
+    every { mockAccountService.currentUserId } returns "test-uid"
+    coEvery { mockUserRepository.getUser("test-uid") } returns testUser
+    coEvery { mockAccountRepository.getAccount("test-uid") } returns testAccount
+    coEvery { mockFeedRepository.getFeedForUids(listOf("test-uid")) } returns testPosts
 
-    coEvery { mockAccountRepository.getAccount("test-uid") } returns
-        Account(
-            uid = "test-uid",
-            ownerId = "test-uid",
-            username = "user1",
-            profilePicture = "",
-            isPrivate = false)
-
-    coEvery { mockLocationRepository.search(any()) } returns emptyList()
-
-    // Mock the fusedLocationClient to avoid lateinit errors
-    LocationProvider.fusedLocationClient = mockk<FusedLocationProviderClient>(relaxed = true)
-
-    viewModel =
-        AccountViewModel(
-            mockAccountService, mockAccountRepository, mockUserRepository, mockLocationRepository)
+    onEditAccountCalled = false
+    onPostClickCalledWithId = null
   }
 
   @After
@@ -102,432 +99,139 @@ class AccountScreenTest {
     clearAllMocks()
   }
 
-  // --- Helpers ---
-  private fun signIn(user: FirebaseUser?) {
-    userFlow.value = user
-    composeTestRule.waitForIdle()
-  }
+  private fun setContent() {
+    viewModel =
+        AccountPageViewModel(
+            mockAccountService, mockAccountRepository, mockUserRepository, mockFeedRepository)
 
-  private fun waitForLoadingToComplete() {
-    composeTestRule.waitUntil(timeoutMillis = 5000) { !viewModel.uiState.value.isLoading }
-    composeTestRule.waitForIdle()
-  }
-
-  private fun setContent(onBack: (() -> Unit)? = null, onSignOut: (() -> Unit)? = null) {
     composeTestRule.setContent {
       OOTDTheme {
-        AccountScreen(
-            accountViewModel = viewModel,
-            credentialManager = mockCredentialManager,
-            onBack = onBack ?: {},
-            onSignOut = onSignOut ?: {})
+        AccountPage(
+            accountModel = viewModel,
+            onEditAccount = { onEditAccountCalled = true },
+            onPostClick = { postId -> onPostClickCalledWithId = postId })
       }
     }
     composeTestRule.waitForIdle()
   }
 
-  private fun selectTestTag(tag: String) = composeTestRule.onNodeWithTag(tag)
-
-  // --- Tests (fewer, but comprehensive) ---
-
   @Test
-  fun rendersBasic_withoutAvatar_andShowsUserInfo_andReadOnlyState() {
-    signIn(mockFirebaseUser)
+  fun screen_displays_user_posts() {
     setContent()
 
-    // Core chrome (back arrow removed from UI; don't assert it)
-    // selectTestTag(UiTestTags.TAG_ACCOUNT_BACK).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_ACCOUNT_TITLE).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_ACCOUNT_AVATAR_CONTAINER).assertIsDisplayed()
+    // Should display all posts (we can check by counting nodes with POST_TAG)
+    composeTestRule
+        .onAllNodesWithTag(AccountPageTestTags.POST_TAG)
+        .fetchSemanticsNodes()
+        .size
+        .let { count -> assert(count == 3) { "Expected 3 posts, found $count" } }
+  }
 
-    // No photo -> letter avatar shown, image absent
-    selectTestTag(UiTestTags.TAG_ACCOUNT_AVATAR_LETTER).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_ACCOUNT_AVATAR_IMAGE).assertDoesNotExist()
+  @Test
+  fun clicking_post_calls_onPostClick_with_correct_id() {
+    setContent()
 
-    // Actions and fields
-    selectTestTag(UiTestTags.TAG_ACCOUNT_EDIT).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).assertIsDisplayed().assertTextContains("user1")
-    selectTestTag(UiTestTags.TAG_GOOGLE_FIELD)
+    // Click on first post
+    composeTestRule.onAllNodesWithTag(AccountPageTestTags.POST_TAG)[0].performClick()
+
+    assert(onPostClickCalledWithId != null) { "onPostClick should have been called" }
+    assert(onPostClickCalledWithId in listOf("post1", "post2", "post3")) {
+      "Post ID should be one of the test posts"
+    }
+  }
+
+  @Test
+  fun screen_displays_loading_indicator_when_loading() {
+    coEvery { mockUserRepository.getUser("test-uid") } coAnswers
+        {
+          kotlinx.coroutines.delay(5000)
+          testUser
+        }
+
+    setContent()
+
+    composeTestRule.onNodeWithTag(AccountPageTestTags.LOADING).assertIsDisplayed()
+  }
+
+  @Test
+  fun screen_handles_empty_posts_list() {
+    coEvery { mockFeedRepository.getFeedForUids(listOf("test-uid")) } returns emptyList()
+
+    setContent()
+
+    // "Your posts :" section should still be displayed
+    composeTestRule.onNodeWithTag(AccountPageTestTags.YOUR_POST_SECTION).assertIsDisplayed()
+
+    // But no post items should exist
+    composeTestRule
+        .onAllNodesWithTag(AccountPageTestTags.POST_TAG)
+        .fetchSemanticsNodes()
+        .size
+        .let { count -> assert(count == 0) { "Expected 0 posts, found $count" } }
+  }
+
+  @Test
+  fun screen_displays_correct_avatar_letter_for_lowercase_username() {
+    coEvery { mockUserRepository.getUser("test-uid") } returns
+        testUser.copy(username = "alice", profilePicture = "")
+    coEvery { mockAccountRepository.getAccount("test-uid") } returns
+        testAccount.copy(username = "alice", profilePicture = "")
+
+    setContent()
+
+    composeTestRule
+        .onNodeWithTag(AccountPageTestTags.AVATAR_LETTER)
         .assertIsDisplayed()
-        .assertTextContains("user1@google.com")
-    selectTestTag(UiTestTags.TAG_SIGNOUT_BUTTON).assertExists()
-
-    // Read-only (not editing)
-    selectTestTag(UiTestTags.TAG_USERNAME_EDIT).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertDoesNotExist()
+        .assertTextContains("A")
   }
 
   @Test
-  fun showsAvatarImage_whenProfilePictureExists() {
-    val avatarUri = Uri.parse("https://example.com/avatar.jpg")
-    every { mockFirebaseUser.photoUrl } returns avatarUri
+  fun screen_handles_username_with_special_characters() {
+    coEvery { mockUserRepository.getUser("test-uid") } returns
+        testUser.copy(username = "@user123", profilePicture = "")
     coEvery { mockAccountRepository.getAccount("test-uid") } returns
-        Account(
-            uid = "test-uid",
-            ownerId = "test-uid",
-            username = "user1",
-            profilePicture = avatarUri.toString())
-
-    signIn(mockFirebaseUser)
-    setContent()
-
-    selectTestTag(UiTestTags.TAG_ACCOUNT_AVATAR_IMAGE).assertIsDisplayed()
-  }
-
-  @Test
-  fun usernameEdit_flow_saveWithoutChange_restores() {
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Open edit
-    selectTestTag(UiTestTags.TAG_USERNAME_EDIT).performClick()
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertIsDisplayed()
-
-    // Save without change -> stays editing (no-op)
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).performClick()
-
-    composeTestRule.waitUntil(
-        condition = { selectTestTag(UiTestTags.TAG_USERNAME_EDIT).isDisplayed() },
-        timeoutMillis = 5000)
-
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertDoesNotExist()
-  }
-
-  @Test
-  fun usernameEdit_save_works() {
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Open edit
-    selectTestTag(UiTestTags.TAG_USERNAME_EDIT).performClick()
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertIsDisplayed()
-
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).performTextInput("new_text")
-
-    // Save with change -> stays editing (no-op)
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).performClick()
-
-    composeTestRule.waitUntil(
-        condition = { selectTestTag(UiTestTags.TAG_USERNAME_EDIT).isDisplayed() },
-        timeoutMillis = 5000)
-
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertDoesNotExist()
-  }
-
-  @Test
-  fun privacyToggle_click_updatesLabel_andCallsRepository() {
-    coEvery { mockAccountRepository.togglePrivacy("test-uid") } returns true
-
-    signIn(mockFirebaseUser)
-    setContent()
-    waitForLoadingToComplete()
-
-    // Wait for the privacy toggle to be displayed
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule.onAllNodesWithText("Public").fetchSemanticsNodes().isNotEmpty()
-    }
-
-    // Scroll to privacy toggle to make it visible on smaller screens
-    selectTestTag(UiTestTags.TAG_PRIVACY_TOGGLE).performScrollTo()
-
-    // Initially Public - use exists instead of assertIsDisplayed for text nodes
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule.onAllNodesWithText("Public").fetchSemanticsNodes().isNotEmpty()
-    }
-
-    val switchMatcher = isToggleable() and hasAnyAncestor(hasTestTag(UiTestTags.TAG_PRIVACY_TOGGLE))
-    composeTestRule.onNode(switchMatcher).performClick()
-    composeTestRule.waitForIdle()
-
-    // Wait for the text to update to Private
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule.onAllNodesWithText("Private").fetchSemanticsNodes().isNotEmpty()
-    }
-
-    // Updated and repo called - verify the text exists in the toggle
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule.onAllNodesWithText("Private").fetchSemanticsNodes().isNotEmpty()
-    }
-    coVerify(exactly = 1) { mockAccountRepository.togglePrivacy("test-uid") }
-  }
-
-  @Test
-  fun helpIcon_showsAndDismissesPopup() {
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Scroll to privacy toggle to make help icon visible on smaller screens
-    selectTestTag(UiTestTags.TAG_PRIVACY_TOGGLE).performScrollTo()
-
-    selectTestTag(UiTestTags.TAG_PRIVACY_HELP_MENU).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_PRIVACY_HELP_ICON).performClick()
-    selectTestTag(UiTestTags.TAG_PRIVACY_HELP_MENU).assertExists()
-    selectTestTag(UiTestTags.TAG_PRIVACY_HELP_MENU).performClick()
-    selectTestTag(UiTestTags.TAG_PRIVACY_HELP_MENU).assertDoesNotExist()
-  }
-
-  @Test
-  fun deleteProfilePicture_buttonAppearsOnlyWithPicture_andCallsViewModel() {
-    // Setup account with profile picture
-    val avatarUri = "https://example.com/avatar.jpg"
-    coEvery { mockAccountRepository.getAccount("test-uid") } returns
-        Account(
-            uid = "test-uid",
-            ownerId = "test-uid",
-            username = "user1",
-            profilePicture = avatarUri,
-            isPrivate = false)
-
-    every { mockAccountService.currentUserId } returns "test-uid"
-    coEvery { mockAccountRepository.deleteProfilePicture("test-uid") } returns Unit
-    coEvery { mockUserRepository.deleteProfilePicture("test-uid") } returns Unit
-
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Delete button should be visible when there's a profile picture
-    selectTestTag(UiTestTags.TAG_ACCOUNT_DELETE).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_ACCOUNT_EDIT).assertIsDisplayed()
-    composeTestRule.onNodeWithText("Edit").assertIsDisplayed()
-
-    // Click delete button
-    selectTestTag(UiTestTags.TAG_ACCOUNT_DELETE).performClick()
-    composeTestRule.waitForIdle()
-
-    // Verify both repositories were called
-    coVerify(exactly = 1) { mockAccountRepository.deleteProfilePicture("test-uid") }
-    coVerify(exactly = 1) { mockUserRepository.deleteProfilePicture("test-uid") }
-  }
-
-  @Test
-  fun deleteProfilePicture_buttonHiddenWhenNoProfilePictureAndShowsUpload() {
-    // Account without profile picture (default setup)
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Delete button should not exist when there's no profile picture
-    selectTestTag(UiTestTags.TAG_ACCOUNT_DELETE).assertDoesNotExist()
-    // Upload button should be shown instead of Edit
-    composeTestRule.onNodeWithText("Upload").assertIsDisplayed()
-  }
-
-  @Test
-  fun handlePickedProfileImage_coversSuccessErrorAndBlank() {
-    val ctx: Context = ApplicationProvider.getApplicationContext()
-
-    // Success: edit is called with expected URL
-    run {
-      var editedCalled = false
-      var editedUrl: String? = null
-      val upload: (String, (String) -> Unit, (Throwable) -> Unit) -> Unit = { _, onSuccess, _ ->
-        onSuccess("https://cdn.example.com/avatar.jpg")
-      }
-
-      composeTestRule.runOnUiThread {
-        handlePickedProfileImage(
-            localPath = "/local/path.jpg",
-            upload = upload,
-            editProfilePicture = {
-              editedCalled = true
-              editedUrl = it
-            },
-            context = ctx)
-      }
-
-      assertTrue(editedCalled)
-      assertTrue(editedUrl == "https://cdn.example.com/avatar.jpg")
-    }
-
-    // Error: edit is not called
-    run {
-      var editedCalled = false
-      val upload: (String, (String) -> Unit, (Throwable) -> Unit) -> Unit = { _, _, onError ->
-        onError(IllegalStateException("upload failed"))
-      }
-
-      composeTestRule.runOnUiThread {
-        handlePickedProfileImage(
-            localPath = "/local/path.jpg",
-            upload = upload,
-            editProfilePicture = { editedCalled = true },
-            context = ctx)
-      }
-
-      assertFalse(editedCalled)
-    }
-
-    // Blank path: no-op (upload not invoked, edit not called)
-    run {
-      var uploadInvoked = false
-      var editedCalled = false
-      val upload: (String, (String) -> Unit, (Throwable) -> Unit) -> Unit = { _, _, _ ->
-        uploadInvoked = true
-      }
-
-      composeTestRule.runOnUiThread {
-        handlePickedProfileImage(
-            localPath = "",
-            upload = upload,
-            editProfilePicture = { editedCalled = true },
-            context = ctx)
-      }
-
-      assertFalse(uploadInvoked)
-      assertFalse(editedCalled)
-    }
-  }
-
-  @Test
-  fun editingUsername_pressEnter_savesAndExitsEditMode() {
-    // Arrange
-    signIn(mockFirebaseUser)
-    every { mockAccountService.currentUserId } returns "test-uid"
-    coEvery { mockAccountRepository.editAccount("test-uid", any(), any(), any(), any()) } returns
-        Unit
-    coEvery { mockUserRepository.editUser("test-uid", any(), any()) } returns Unit
+        testAccount.copy(username = "@user123", profilePicture = "")
 
     setContent()
 
-    // Enter edit mode
-    selectTestTag(UiTestTags.TAG_USERNAME_EDIT).performClick()
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertIsDisplayed()
-
-    // Replace the text explicitly to avoid caret position issues, then press IME Done
-    val newUsername = "user1X"
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).performTextClearance()
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).performTextInput(newUsername)
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).performImeAction()
-    composeTestRule.waitForIdle()
-
-    // Assert: edit controls gone, field shows updated username
-    selectTestTag(UiTestTags.TAG_USERNAME_EDIT).assertIsDisplayed()
-    selectTestTag(UiTestTags.TAG_USERNAME_CANCEL).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_USERNAME_SAVE).assertDoesNotExist()
-    selectTestTag(UiTestTags.TAG_USERNAME_FIELD).assertTextContains(newUsername)
-
-    // Verify repository updates invoked with new username (location param included)
-    coVerify { mockAccountRepository.editAccount("test-uid", newUsername, any(), any(), any()) }
-    coVerify { mockUserRepository.editUser("test-uid", newUsername, any()) }
-  }
-
-  // --- Location field tests ---
-
-  @Test
-  fun locationField_displaysSavedLocation_whenAccountHasLocation() {
-    // Setup account with saved location
-    val savedLocation = Location(46.5191, 6.5668, "EPFL, Lausanne")
-    coEvery { mockAccountRepository.getAccount("test-uid") } returns
-        Account(
-            uid = "test-uid",
-            ownerId = "test-uid",
-            username = "user1",
-            profilePicture = "",
-            isPrivate = false,
-            location = savedLocation)
-
-    signIn(mockFirebaseUser)
-    setContent()
-    waitForLoadingToComplete()
-
-    // Location field should display the saved location - use the correct test tag
-    selectTestTag(LocationSelectionTestTags.INPUT_LOCATION)
+    composeTestRule
+        .onNodeWithTag(AccountPageTestTags.USERNAME_TEXT)
         .assertIsDisplayed()
-        .assertTextContains("EPFL, Lausanne")
+        .assertTextContains("@user123")
   }
 
   @Test
-  fun locationField_clearButton_makesFieldEditable() {
-    // Setup account with saved location
-    val savedLocation = Location(46.5191, 6.5668, "EPFL, Lausanne")
-    coEvery { mockAccountRepository.getAccount("test-uid") } returns
-        Account(
-            uid = "test-uid",
-            ownerId = "test-uid",
-            username = "user1",
-            profilePicture = "",
-            isPrivate = false,
-            location = savedLocation)
-
-    signIn(mockFirebaseUser)
+  fun screen_structure_is_complete() {
     setContent()
-    waitForLoadingToComplete()
 
-    // Verify location is displayed - use the correct test tag
-    selectTestTag(LocationSelectionTestTags.INPUT_LOCATION).assertTextContains("EPFL, Lausanne")
-
-    // Click clear button
-    selectTestTag(LocationSelectionTestTags.LOCATION_CLEAR_BUTTON).performClick()
-    composeTestRule.waitForIdle()
-
-    // Field should now be editable (text cleared)
-    selectTestTag(LocationSelectionTestTags.INPUT_LOCATION).performTextInput("Zurich")
-    composeTestRule.waitForIdle()
-
-    // Verify text was entered (field is editable)
-    selectTestTag(LocationSelectionTestTags.INPUT_LOCATION).assertTextContains("Zurich")
+    // Verify all main components are present
+    composeTestRule.onNodeWithTag(AccountPageTestTags.TITLE_TEXT).assertIsDisplayed()
+    composeTestRule.onNodeWithTag(AccountPageTestTags.SETTINGS_BUTTON).assertIsDisplayed()
+    composeTestRule.onNodeWithTag(AccountPageTestTags.USERNAME_TEXT).assertIsDisplayed()
+    composeTestRule.onNodeWithTag(AccountPageTestTags.FRIEND_COUNT_TEXT).assertIsDisplayed()
+    composeTestRule.onNodeWithTag(AccountPageTestTags.YOUR_POST_SECTION).assertIsDisplayed()
   }
 
   @Test
-  fun locationField_gpsButton_exists() {
-    signIn(mockFirebaseUser)
+  fun screen_handles_large_number_of_posts() {
+    val manyPosts =
+        (1..20).map { index ->
+          OutfitPost(
+              postUID = "post$index",
+              name = "testuser",
+              ownerId = "test-uid",
+              outfitURL = "https://example.com/post$index.jpg",
+              description = "Post $index")
+        }
+
+    coEvery { mockFeedRepository.getFeedForUids(listOf("test-uid")) } returns manyPosts
+
     setContent()
 
-    // GPS button should be visible
-    selectTestTag(LocationSelectionTestTags.LOCATION_GPS_BUTTON).assertIsDisplayed()
-    composeTestRule.onNodeWithText("Update Location (GPS)").assertIsDisplayed()
-  }
-
-  // --- EPFL Default Location tests ---
-
-  @Test
-  fun defaultEpflLocationButton_isDisplayed() {
-    signIn(mockFirebaseUser)
-    setContent()
-
-    // Wait for the EPFL button to exist in the semantics tree
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule
-          .onAllNodesWithText("or select default location (EPFL)", substring = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-
-    selectTestTag(LocationSelectionTestTags.LOCATION_DEFAULT_EPFL)
-        .performScrollTo()
-        .assertIsDisplayed()
-        .assertTextContains("or select default location (EPFL)")
-  }
-
-  @Test
-  fun defaultEpflLocationButton_setsLocationToEpfl() {
-    signIn(mockFirebaseUser)
-    setContent()
-    waitForLoadingToComplete()
-
-    // Wait for the EPFL button to exist in the semantics tree
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule
-          .onAllNodesWithText("or select default location (EPFL)", substring = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-    // Scroll to make the button visible on smaller screens
-    selectTestTag(LocationSelectionTestTags.LOCATION_DEFAULT_EPFL).performScrollTo().performClick()
-    composeTestRule.waitForIdle()
-
-    // Assert: location should be set to EPFL
-    composeTestRule.runOnUiThread {
-      val selectedLocation = viewModel.uiState.value.location
-      assertEquals(46.5191, selectedLocation.latitude, 0.0001)
-      assertEquals(6.5668, selectedLocation.longitude, 0.0001)
-      assertTrue(selectedLocation.name.contains("EPFL"))
-      assertEquals(selectedLocation.name, viewModel.uiState.value.locationQuery)
-    }
+    composeTestRule
+        .onAllNodesWithTag(AccountPageTestTags.POST_TAG)
+        .fetchSemanticsNodes()
+        .size
+        .let { count -> assert(count == 20) { "Expected 20 posts, found $count" } }
   }
 }
