+++ app/src/androidTest/java/com/android/ootd/ui/account/EditAccountViewModelTest.kt
@@ -44,7 +44,7 @@ import org.junit.Test
  */
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class AccountViewModelTest {
+class EditAccountViewModelTest {
 
   private lateinit var accountService: AccountService
   private lateinit var accountRepository: AccountRepository
